P23 release for bcm4398, 2022.09.23 Version : 103.10.110.1
[Features]
Concurrent features (Dual STA, Dual AP, STA+SAP)

[Bug]
BUG:241207159 Support TxPwr CAP Blob for WLAN SAR Limits(SAR Blob)
BUG:242295231 Fixed Station Dump Info Missing
BUG:242991260 Fixed SoftAP ACS failed Issue
BUG:237486831 Optimized 30ms gap during assoc and probs
BUG:240243201,BUG:238864408 Support PCIe loopback for 4398c0
BUG:243334800 Fixed to use randomized source address for ANQP frame
BUG:242167842 Fixed TDLS Features Capability missing
BUG:242876059 Fixed random phy trap during connect
BUG:242848352 Fixed to support OFDM Scrambler
BUG:243916098 Fixed the DSCP map setting command
BUG:243887033 Fixed Kernel Warning
BUG:234666930 Feature to measure RSSI
BUG:243713728 High PER Hump/Spike Due To CW-Tone At Low Level

Bug: 248430015
Test: SVT release test

+++ include/event_log_payload.h
@@ -1507,12 +1507,12 @@ typedef struct wlc_rx_fifo_overflow_info_v1 {
 
 	uint32 rxfill[WLC_RX_FIFO_DMA_NUM];
 	uint32 dma_stall_check_rxfill[WLC_RX_FIFO_DMA_NUM];
-	uint32 rx_dma_fill_success_time[WLC_RX_FIFO_DMA_NUM];
-	uint32 rx_dma_fill_fail_time[WLC_RX_FIFO_DMA_NUM];
+	uint64 rx_dma_fill_success_time[WLC_RX_FIFO_DMA_NUM];	/* in ns */
+	uint64 rx_dma_fill_fail_time[WLC_RX_FIFO_DMA_NUM];	/* in ns */
 
 	uint32 rxposts[WLC_RX_FIFO_DMA_NUM];
 	uint32 dma_stall_check_rxposts[WLC_RX_FIFO_DMA_NUM];
-	uint32 rx_dma_posts_success_time[WLC_RX_FIFO_DMA_NUM];
+	uint64 rx_dma_posts_success_time[WLC_RX_FIFO_DMA_NUM];	/* in ns */
 
 	uint32 rx_dma_desc_count[WLC_RX_FIFO_DMA_NUM];
 } wlc_rx_fifo_overflow_info_v1_t;
